change: move CreeAndEnglish to search.py
